requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
billiard==4.1.0
bio==1.5.9
biopython==1.81
biothings-client==0.3.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.2
django-celery-results==2.5.1
django-cors-headers==4.1.0
djangorestframework==3.14.0
gprofiler-official==1.0.0
idna==3.4
kombu==5.3.1
mygene==3.2.2
numpy==1.25.0
packaging==23.1
pandas==2.0.2
platformdirs==3.8.0
pooch==1.7.0
prompt-toolkit==3.0.38
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.65.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
